docs(carousel): add indicators example
