+++ packages/oruga/src/components/carousel/examples/autoplay.vue
@@ -40,6 +40,7 @@ const repeat = ref(false);
                     label="Repeat" />
             </o-field>
         </o-field>
+
         <o-field grouped>
             <o-field label="Value">
                 <o-input
@@ -66,7 +67,8 @@ const repeat = ref(false);
             :autoplay="autoplay"
             :pause-hover="pauseHover"
             :interval="interval"
-            :repeat="repeat">
+            :repeat="repeat"
+            :arrows-hover="false">
             <o-carousel-item v-for="(slide, i) in carousels" :key="i">
                 <section
                     class="example-slide"
